New packet counter (unsigned long) which is used as parameter to dataHandler() callback. It counts in parallel with old one but unlike it, new one can go without rollover. The old one is still in use to set correct packet header values.
XModem class tested on Arduino Uno as transmitter. It is kind of
slow - with 115200bps it can barely reach 4100cps. At least it
is working.

+++ test/AllTests.cpp
@@ -22,12 +22,12 @@ void sendChar(char sym)
 {
 	mock().actualCall("sendChar").withParameter("sym", sym);
 }
-bool dataHandler(int no, char* data, int size)
+bool dataHandler(unsigned long no, char* data, int size)
 {
 	if (mock().hasData("dataHandler"))
 	{	if(mock().getData("dataHandler").getIntValue() == 1)
 			return (mock().actualCall("dataHandler").
-				withParameter("no", no).
+				withParameter("no", (int)no).
 				withParameter("data", data[0]).
 				withParameter("size", size).
 				returnValue().getIntValue() == 1)?
@@ -35,7 +35,7 @@ bool dataHandler(int no, char* data, int size)
 	
 		if(mock().getData("dataHandler").getIntValue() == 2)
 			return (mock().actualCall("dataHandler").
-				withParameter("no", no).
+				withParameter("no", (int)no).
 				withParameter("size", size).
 				returnValue().getIntValue() == 1)?
 				true:false;						
@@ -44,7 +44,7 @@ bool dataHandler(int no, char* data, int size)
 		{
 			memset((void*)data, mock().getData("dataHandler_data").getIntValue(), 128);	
 			return (mock().actualCall("dataHandler").
-				withParameter("no", no).
+				withParameter("no", (int)no).
 				withParameter("size", size).
 				returnValue().getIntValue() == 1)?
 				true:false;
